Add external css for entire website
I also added a description below the picture and title for homemade egg
sandwich on the main webpage. I tried to add a button to link to the
webpage for the egg sandwich, but I don't know how yet. I then tried to
improvise with a div and margins to make a link look like a button, but
that hasn't worked out yet either.