Start styling entire website and start 2nd recipes
I started filling out the webpage for the eggs and home fries recipe. I
also started to fill out the CSS so that the website doesn't look like
doo-doo. I gave all the webpages a custom font, font color, and
background color. I changed the margins so that the content of each
webpage is more centered, and I gave each recipe description on the
main webpage a set width constraint. Finally, I fixed the heights of
the images.